Simplify audio player
