+++ src-tauri/src/models.rs
@@ -1,0 +1,3 @@
+mod audio_request;
+
+pub use audio_request::*;
