+++ src-tauri/src/services.rs
@@ -1,5 +1,3 @@
-mod audio_service;
-mod download_audio;
+mod audio;
 
-pub use audio_service::*;
-pub use download_audio::*;
+pub use audio::*;
